+++ boringssl/bindgen.h
@@ -13,6 +13,7 @@
 #include <openssl/err.h>
 #include <openssl/evp.h>
 #include <openssl/hmac.h>
+#include <openssl/md5.h>
 #include <openssl/mem.h>
 #include <openssl/rand.h>
 #include <openssl/rsa.h>