+++ src/hash.rs
@@ -103,6 +103,28 @@ pub trait Digest: Eq + PartialEq + Display + Debug + Sized + self::inner::Digest
     fn bytes(&self) -> Self::Bytes;
 }
 
+#[cfg(feature = "insecure")]
+#[deprecated(note = "MD5 is considered insecure")]
+#[allow(deprecated)] // Work-around until Rust issue #56195 is resolved
+#[derive(Clone, Default)]
+pub(crate) struct InsecureMd5 {
+    ctx: CStackWrapper<boringssl::MD5_CTX>,
+}
+
+#[cfg(feature = "insecure")]
+pub(crate) mod insecure_md5_digest {
+    use boringssl;
+
+    /// INSECURE: The digest output by the MD5 hash function.
+    ///
+    /// # Security
+    ///
+    /// MD5 is considered insecure, and should only be used for compatibility
+    /// with legacy applications.
+    #[deprecated(note = "MD5 is considered insecure")]
+    pub struct InsecureMd5Digest(pub(crate) [u8; boringssl::MD5_DIGEST_LENGTH as usize]);
+}
+
 // NOTE: InsecureSha1 is not publicly available; it is only used in HMAC-SHA1.
 #[cfg(feature = "insecure")]
 #[deprecated(note = "SHA-1 is considered insecure")]
@@ -278,6 +300,16 @@ macro_rules! impl_hash {
     };
 }
 
+#[cfg(feature = "insecure")]
+impl_hash!(
+    InsecureMd5,
+    self::insecure_md5_digest::InsecureMd5Digest,
+    boringssl::MD5_DIGEST_LENGTH,
+    md5_update,
+    md5_final,
+    evp_md5,
+    NID_md5
+);
 #[cfg(feature = "insecure")]
 impl_hash!(
     InsecureSha1,