+++ CHANGELOG.md
@@ -10,6 +10,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+- Exposed MD5 digest and HMAC in the `insecure` module.
+
 ## [0.4.2] - 2019-10-04
 
 ### Fixed